requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
aiohappyeyeballs==2.4.3
    # via
    #   -r requirements.in
    #   aiohttp
aiohttp==3.11.7
    # via
    #   -r requirements.in
    #   datasets
    #   fsspec
aiosignal==1.3.1
    # via
    #   -r requirements.in
    #   aiohttp
appnope==0.1.4
    # via
    #   -r requirements.in
    #   ipykernel
asttokens==2.4.1
    # via
    #   -r requirements.in
    #   stack-data
attrs==24.2.0
    # via
    #   -r requirements.in
    #   aiohttp
certifi==2024.8.30
    # via
    #   -r requirements.in
    #   requests
charset-normalizer==3.4.0
    # via
    #   -r requirements.in
    #   requests
comm==0.2.2
    # via
    #   -r requirements.in
    #   ipykernel
    #   ipywidgets
datasets==3.1.0
    # via -r requirements.in
debugpy==1.8.9
    # via
    #   -r requirements.in
    #   ipykernel
decorator==5.1.1
    # via
    #   -r requirements.in
    #   ipython
dill==0.3.8
    # via
    #   -r requirements.in
    #   datasets
    #   multiprocess
executing==2.1.0
    # via
    #   -r requirements.in
    #   stack-data
filelock==3.16.1
    # via
    #   -r requirements.in
    #   datasets
    #   huggingface-hub
    #   torch
    #   transformers
frozenlist==1.5.0
    # via
    #   -r requirements.in
    #   aiohttp
    #   aiosignal
fsspec[http]==2024.9.0
    # via
    #   -r requirements.in
    #   datasets
    #   huggingface-hub
    #   torch
huggingface-hub==0.26.2
    # via
    #   -r requirements.in
    #   datasets
    #   tokenizers
    #   transformers
idna==3.10
    # via
    #   -r requirements.in
    #   requests
    #   yarl
ipykernel==6.29.5
    # via -r requirements.in
ipython==8.29.0
    # via
    #   -r requirements.in
    #   ipykernel
    #   ipywidgets
ipywidgets==8.1.5
    # via -r requirements.in
jedi==0.19.2
    # via
    #   -r requirements.in
    #   ipython
jinja2==3.1.4
    # via
    #   -r requirements.in
    #   torch
jupyter-client==8.6.3
    # via
    #   -r requirements.in
    #   ipykernel
jupyter-core==5.7.2
    # via
    #   -r requirements.in
    #   ipykernel
    #   jupyter-client
jupyterlab-widgets==3.0.13
    # via
    #   -r requirements.in
    #   ipywidgets
markupsafe==3.0.2
    # via
    #   -r requirements.in
    #   jinja2
matplotlib-inline==0.1.7
    # via
    #   -r requirements.in
    #   ipykernel
    #   ipython
mpmath==1.3.0
    # via
    #   -r requirements.in
    #   sympy
multidict==6.1.0
    # via
    #   -r requirements.in
    #   aiohttp
    #   yarl
multiprocess==0.70.16
    # via
    #   -r requirements.in
    #   datasets
nest-asyncio==1.6.0
    # via
    #   -r requirements.in
    #   ipykernel
networkx==3.4.2
    # via
    #   -r requirements.in
    #   torch
numpy==2.1.3
    # via
    #   -r requirements.in
    #   datasets
    #   pandas
    #   torchvision
    #   transformers
packaging==24.2
    # via
    #   -r requirements.in
    #   datasets
    #   huggingface-hub
    #   ipykernel
    #   transformers
pandas==2.2.3
    # via
    #   -r requirements.in
    #   datasets
parso==0.8.4
    # via
    #   -r requirements.in
    #   jedi
pexpect==4.9.0
    # via
    #   -r requirements.in
    #   ipython
pillow==11.0.0
    # via
    #   -r requirements.in
    #   torchvision
platformdirs==4.3.6
    # via
    #   -r requirements.in
    #   jupyter-core
prompt-toolkit==3.0.48
    # via
    #   -r requirements.in
    #   ipython
propcache==0.2.0
    # via
    #   -r requirements.in
    #   aiohttp
    #   yarl
psutil==6.1.0
    # via
    #   -r requirements.in
    #   ipykernel
ptyprocess==0.7.0
    # via
    #   -r requirements.in
    #   pexpect
pure-eval==0.2.3
    # via
    #   -r requirements.in
    #   stack-data
pyarrow==18.1.0
    # via
    #   -r requirements.in
    #   datasets
pygments==2.18.0
    # via
    #   -r requirements.in
    #   ipython
python-dateutil==2.9.0.post0
    # via
    #   -r requirements.in
    #   jupyter-client
    #   pandas
pytz==2024.2
    # via
    #   -r requirements.in
    #   pandas
pyyaml==6.0.2
    # via
    #   -r requirements.in
    #   datasets
    #   huggingface-hub
    #   transformers
pyzmq==26.2.0
    # via
    #   -r requirements.in
    #   ipykernel
    #   jupyter-client
regex==2024.11.6
    # via
    #   -r requirements.in
    #   transformers
requests==2.32.3
    # via
    #   -r requirements.in
    #   datasets
    #   huggingface-hub
    #   transformers
safetensors==0.4.5
    # via
    #   -r requirements.in
    #   transformers
six==1.16.0
    # via
    #   -r requirements.in
    #   asttokens
    #   python-dateutil
stack-data==0.6.3
    # via
    #   -r requirements.in
    #   ipython
sympy==1.13.1
    # via
    #   -r requirements.in
    #   torch
tokenizers==0.20.3
    # via
    #   -r requirements.in
    #   transformers
torch==2.5.1
    # via
    #   -r requirements.in
    #   torchvision
torchvision==0.20.1
    # via -r requirements.in
tornado==6.4.2
    # via
    #   -r requirements.in
    #   ipykernel
    #   jupyter-client
tqdm==4.67.1
    # via
    #   -r requirements.in
    #   datasets
    #   huggingface-hub
    #   transformers
traitlets==5.14.3
    # via
    #   -r requirements.in
    #   comm
    #   ipykernel
    #   ipython
    #   ipywidgets
    #   jupyter-client
    #   jupyter-core
    #   matplotlib-inline
transformers==4.46.3
    # via -r requirements.in
typing-extensions==4.12.2
    # via
    #   -r requirements.in
    #   huggingface-hub
    #   ipython
    #   torch
tzdata==2024.2
    # via
    #   -r requirements.in
    #   pandas
urllib3==2.2.3
    # via
    #   -r requirements.in
    #   requests
wcwidth==0.2.13
    # via
    #   -r requirements.in
    #   prompt-toolkit
widgetsnbextension==4.0.13
    # via
    #   -r requirements.in
    #   ipywidgets
xxhash==3.5.0
    # via
    #   -r requirements.in
    #   datasets
yarl==1.18.0
    # via
    #   -r requirements.in
    #   aiohttp
